Update getecho.py
We tested this code and made necessary modifications in  IUAC Lab (May first week 2014).
Few attempts failed. the problem detected was that of sensor( Maight have happened due to wrong connections). There was no consistency in the readings. Tested with another srf module. and it worked. ( Need to procure new srf05 sensors)